Feature : Connection : Add connnection manager to check connection before making server call

+++ core/repository/src/main/kotlin/com/demo/datausage/core/repository/datausage/DataUsageRepository.kt
@@ -1,8 +1,9 @@
-package com.demo.datausage.core.repository
+package com.demo.datausage.core.repository.datausage
 
 import android.util.Log
 import com.demo.datausage.core.data.datasource.local.database.dao.DataUsageDao
 import com.demo.datausage.core.data.datasource.remote.DataUsageAPI
+import com.demo.datausage.core.repository.connection.ConnectionManager
 import com.demo.datausage.core.repository.mappers.DBDataMapper
 import com.demo.datausage.core.repository.mappers.QuarterWiseDataMapper
 import com.demo.datausage.core.repository.mappers.YearWiseDataMapper
@@ -21,7 +22,8 @@ interface DataUsageRepository {
 
 class DataUsageRepository_Impl(
     private val service : DataUsageAPI,
-    private val dataUsageDao:DataUsageDao
+    private val dataUsageDao:DataUsageDao,
+    private val connectionManager: ConnectionManager
 ) : DataUsageRepository{
     private val resource_id = "a807b7ab-6cad-4aa6-87d0-e283a7353a0f"
     override fun getYearWiseData(): Flow<List<YearWiseData>> = flow{
@@ -29,36 +31,51 @@ class DataUsageRepository_Impl(
         val dbData = dataUsageDao
             .fetchAllData()
             .first()
-
         if(dbData.isNotEmpty()) {
             emit(YearWiseDataMapper.mapToYearWiseDataList(dbData))
         }
-        val response=service.getDataUsage(resource_id)
-        Log.d("DEMO","$response")
-        val dbDataListFromResponse = DBDataMapper.mapFromResponse(response)
-        if(dbDataListFromResponse.isNotEmpty()){
-            dataUsageDao.deleteAll()
-            dataUsageDao.insertAll(dbDataListFromResponse)
+        when(connectionManager.isNetworkAvailable()){
+            true -> {
+                val response=service.getDataUsage(resource_id)
+                Log.d("DEMO","$response")
+                val dbDataListFromResponse = DBDataMapper.mapFromResponse(response)
+                if(dbDataListFromResponse.isNotEmpty()){
+                    dataUsageDao.deleteAll()
+                    dataUsageDao.insertAll(dbDataListFromResponse)
+                }
+                emit(YearWiseDataMapper.mapToYearWiseDataList(dbData))
+            }
+            false -> {
+
+            }
+
         }
-        emit(YearWiseDataMapper.mapToYearWiseDataList(dbData))
+
     }
 
     override fun getQtrWiseData(): Flow<List<QuarterWiseData>> = flow {
         val dbData = dataUsageDao
             .fetchAllData()
             .first()
-
         if(dbData.isNotEmpty()) {
             emit(QuarterWiseDataMapper.mapToQuarterWiseDataList(dbData))
         }
-        val response=service.getDataUsage(resource_id)
-        Log.d("DEMO","$response")
-        val dbDataListFromResponse = DBDataMapper.mapFromResponse(response)
-        if(dbDataListFromResponse.isNotEmpty()){
-            dataUsageDao.deleteAll()
-            dataUsageDao.insertAll(dbDataListFromResponse)
+        when(connectionManager.isNetworkAvailable()){
+            true -> {
+                val response=service.getDataUsage(resource_id)
+                Log.d("DEMO","$response")
+                val dbDataListFromResponse = DBDataMapper.mapFromResponse(response)
+                if(dbDataListFromResponse.isNotEmpty()){
+                    dataUsageDao.deleteAll()
+                    dataUsageDao.insertAll(dbDataListFromResponse)
+                }
+                emit(QuarterWiseDataMapper.mapToQuarterWiseDataList(dbData))
+            }
+            false -> {
+
+            }
         }
-        emit(QuarterWiseDataMapper.mapToQuarterWiseDataList(dbData))
+
 
     }
 }
